perf(workflows): lazy-load @xyflow/react WorkflowGraph and scope its CSS
Splits WorkflowGraph.tsx into:
  - WorkflowGraphImpl.tsx — owns every @xyflow/react runtime import + the
    @xyflow/react/dist/style.css side-effect import. Default-exports the
    ReactFlow component.
  - WorkflowGraph.tsx — thin client wrapper. Loads the Impl via
    next/dynamic({ ssr: false }) and forwards props. Re-exports
    WorkflowGraphReadOnly as a thin convenience wrapper.

Removes the global @xyflow/react/dist/style.css import from
apps/mercato/src/app/globals.css; the stylesheet now ships in the lazy
chunk next to the component that needs it. Every route other than the
workflows visual editor / instance viewer no longer pays the xyflow
parse + style cost.

Type-only imports of Node / Edge / Connection in the wrapper are fully
erased at runtime, so no @xyflow/react code enters the wrapper's
chunk.
