+++ .ai/runs/2026-05-27-dev-mode-lazy-load-heavy-clients/PLAN.md
@@ -16,8 +16,8 @@
 | Phase | Step | Title | Status | Commit |
 |-------|------|-------|--------|--------|
 | 1 | 1.1 | Seed run folder (PLAN, HANDOFF, NOTIFY) | done | (this commit) |
-| 2 | 2.1 | Lazy-load recharts chart primitives via next/dynamic | done | pending |
-| 2 | 2.2 | Lazy-load @xyflow/react WorkflowGraph and scope its CSS | todo | — |
+| 2 | 2.1 | Lazy-load recharts chart primitives via next/dynamic | done | 75877478e |
+| 2 | 2.2 | Lazy-load @xyflow/react WorkflowGraph and scope its CSS | done | pending |
 | 2 | 2.3 | Defer ClientBootstrap registry barrels to client-side mount | todo | — |
 | 3 | 3.1 | Enable Next.js optimizePackageImports for lucide-react, recharts, date-fns | todo | — |
 | 3 | 3.2 | Remove dead transpiledWorkspacePackages computation | todo | — |
